make solidityscan report component reusable
